add update password and email to UserService

+++ back/src/main/java/com/openclassrooms/mdd/repository/ThemeRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.mdd.repository;
 import com.openclassrooms.mdd.model.Theme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThemeRepository extends JpaRepository<Theme, Long> {
+public interface ThemeRepository extends JpaRepository<Theme, Integer> {
 }
